+++ tests/attributes/nacc/modules/uds/test_naccnihr.py
@@ -1,21 +1,13 @@
-"""Tests create_naccnihr specifically."""
+"""Tests create_race specifically."""
 
-from typing import Callable
-
-import pytest
-from nacc_attribute_deriver.attributes.nacc.modules.uds.form_a1 import (
-    UDSFormA1Attribute,
+from nacc_attribute_deriver.attributes.nacc.modules.uds.helpers.generate_race import (
+    generate_race,
 )
 
 
-@pytest.fixture(scope="module")
-def generate_naccnihr() -> Callable:
-    return UDSFormA1Attribute.generate_naccnihr
-
-
-class TestCreateNACCNIHR:
-    """Specifically test the generate_naccnihr function, which ultimately is
-    also testing _create_naccnihr. Needs.
+class TestCreateRace:
+    """Specifically test the generate_race function, which ultimately is also
+    testing _create_naccnihr and _create_naccninr. Needs.
 
     race
     racex
@@ -25,58 +17,58 @@ class TestCreateNACCNIHR:
     raceterx
     """
 
-    def test_original_primary(self, generate_naccnihr):
-        assert generate_naccnihr(None, None, None, None, None, None) == 99
-        assert generate_naccnihr(1, None, None, None, None, None) == 1
-        assert generate_naccnihr(2, None, None, None, None, None) == 2
-        assert generate_naccnihr(3, None, None, None, None, None) == 3
-        assert generate_naccnihr(4, None, None, None, None, None) == 4
-        assert generate_naccnihr(5, None, None, None, None, None) == 5
-        assert generate_naccnihr(6, None, None, None, None, None) == 6
-        assert generate_naccnihr(88, None, None, None, None, None) == 88
+    def test_original_primary(self):
+        assert generate_race(None, None, None, None, None, None) == 99
+        assert generate_race(1, None, None, None, None, None) == 1
+        assert generate_race(2, None, None, None, None, None) == 2
+        assert generate_race(3, None, None, None, None, None) == 3
+        assert generate_race(4, None, None, None, None, None) == 4
+        assert generate_race(5, None, None, None, None, None) == 5
+        assert generate_race(6, None, None, None, None, None) == 6
+        assert generate_race(88, None, None, None, None, None) == 88
 
-    def test_original_primary_writein(self, generate_naccnihr):
-        assert generate_naccnihr(50, "Arab", None, None, None, None) == 1
-        assert generate_naccnihr(50, "African American", None, None, None, None) == 2
-        assert generate_naccnihr(50, "NATIVE AMERICAN", None, None, None, None) == 3
+    def test_original_primary_writein(self):
+        assert generate_race(50, "Arab", None, None, None, None) == 1
+        assert generate_race(50, "African American", None, None, None, None) == 2
+        assert generate_race(50, "NATIVE AMERICAN", None, None, None, None) == 3
 
         # TODO: check what SAS/R code returns for this
         # won't return 4
-        # assert generate_naccnihr(50, "Samoan", None, None, None, None) == 4
+        # assert generate_race(50, "Samoan", None, None, None, None) == 4
         # this version will return 4 (must be racesecx)
-        assert generate_naccnihr(50, None, None, "Samoan", None, None) == 4
+        assert generate_race(50, None, None, "Samoan", None, None) == 4
 
-        assert generate_naccnihr(50, "Tahitian", None, None, None, None) == 4
+        assert generate_race(50, "Tahitian", None, None, None, None) == 4
 
-        assert generate_naccnihr(50, "Asian", None, None, None, None) == 5
-        assert generate_naccnihr(50, "Biracial", None, None, None, None) == 6
+        assert generate_race(50, "Asian", None, None, None, None) == 5
+        assert generate_race(50, "Biracial", None, None, None, None) == 6
         assert (
-            generate_naccnihr(50, "African and American Indian", None, None, None, None)
+            generate_race(50, "African and American Indian", None, None, None, None)
             == 6
         )
-        assert generate_naccnihr(50, "HUMAN", None, None, None, None) == 99
+        assert generate_race(50, "HUMAN", None, None, None, None) == 99
 
-    def test_original_ignore(self, generate_naccnihr):
+    def test_original_ignore(self):
         # seems like it should ignore the racex
-        assert generate_naccnihr(1, "Arab", None, None, None, None) == 1
-        assert generate_naccnihr(2, "Arab", None, None, None, None) == 2
-        assert generate_naccnihr(3, "Arab", None, None, None, None) == 3
-        assert generate_naccnihr(4, "Arab", None, None, None, None) == 4
-        assert generate_naccnihr(5, "Arab", None, None, None, None) == 6
-        assert generate_naccnihr(6, "Arab", None, None, None, None) == 6
+        assert generate_race(1, "Arab", None, None, None, None) == 1
+        assert generate_race(2, "Arab", None, None, None, None) == 2
+        assert generate_race(3, "Arab", None, None, None, None) == 3
+        assert generate_race(4, "Arab", None, None, None, None) == 4
+        assert generate_race(5, "Arab", None, None, None, None) == 6
+        assert generate_race(6, "Arab", None, None, None, None) == 6
 
     # the following are pulled from regression testing
 
-    def test_NACC359394(self, generate_naccnihr):
+    def test_NACC359394(self):
         # baseline says 99, computed seems more correct
-        assert generate_naccnihr(50, "Mulato", 88, "", 88, "") == 6
+        assert generate_race(50, "Mulato", 88, "", 88, "") == 6
 
-    def test_NACC201235(self, generate_naccnihr):
+    def test_NACC201235(self):
         # baseline says 1, was incorrectly computing as 6
         # was missing itialian american from white_responses, fixed
-        assert generate_naccnihr(1, "", 50, "ITIALIAN AMERICAN", 88, "") == 1
+        assert generate_race(1, "", 50, "ITIALIAN AMERICAN", 88, "") == 1
 
-    def test_NACC356772(self, generate_naccnihr):
+    def test_NACC356772(self):
         """baseline says 6, was computing as 99 EGYPT as a response is not
         defined in the SAS code, I believe it's set by the line.
 
@@ -88,23 +80,23 @@ class TestCreateNACCNIHR:
               then &NACCNIHR = 99;
         updated code to work on if/else similar to SAS
         """
-        assert generate_naccnihr(50, "EGYPT", 1, "", 88, "") == 6
+        assert generate_race(50, "EGYPT", 1, "", 88, "") == 6
 
-    def test_NACC703416(self, generate_naccnihr):
+    def test_NACC703416(self):
         # baseline says 1, computed says 6
         # also affected by the if/else inconsistency, fixed now
-        assert generate_naccnihr(50, "POLISH", 1, "", 88, "") == 1
+        assert generate_race(50, "POLISH", 1, "", 88, "") == 1
 
-    def test_hispanic(self, generate_naccnihr):
+    def test_hispanic(self):
         """This one is most volatile to changes/script differences.
 
         Original derived value is 99, but needs to be further refined
         for UDSv4
         """
-        assert generate_naccnihr(50, "HISPANIC", 88, "", 88, "") == 99
-        assert generate_naccnihr(50, "Hispanic", 88, "", 88, "") == 99
+        assert generate_race(50, "HISPANIC", 88, "", 88, "") == 99
+        assert generate_race(50, "Hispanic", 88, "", 88, "") == 99
 
-    def test_NACC342334(self, generate_naccnihr):
+    def test_NACC342334(self):
         """baseline 1 vs computed 6, issue was accidental lack of comma should
         in fact be 1."""
-        assert generate_naccnihr(1, "", 50, "Middle Eastern", 88, "") == 1
+        assert generate_race(1, "", 50, "Middle Eastern", 88, "") == 1