+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Documentation of release versions of `nacc-attribute-deriver`
 * Adds ability to derive type returned by a curation rule with tool to create table of rule types.
 * Adds enforcement of expected types when grabbing variables from the raw data
 * Adds rest of NP variables
-* Adds rest of UDS A1 derived variables
+* Adds rest of UDS A1, A2 derived variables
+
 
 ## 1.4.0